src/backend/pyproject.toml
[tool.poetry]
name = "rag-backend"
version = "0.1.0"
description = ""
authors = ["DY Lim <dlagus6412@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11,<3.13"
beautifulsoup4 = "4.12.3"
flask = {extras = ["async"], version = "^3.1.0"}
langchain = "0.3.7"
langchain-community = "0.3.7"
langchain-core = "0.3.19"
langchain-google-vertexai = "2.0.7"
langchain-qdrant = "0.2.0"
langchain-text-splitters = "0.3.2"
langgraph = "0.2.50"
pyyaml = "6.0.2"
qdrant-client = "1.12.1"
typing-extensions = "4.12.2"
gunicorn = "^23.0.0"
pypdf = "4.3.1"
aiohttp = "3.11.5"
langchain-upstage = "^0.4.0"
pymupdf = "^1.24.14"
pillow = "^11.0.0"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pipreqs = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
